--- Runtime/Systems/AbilityUpdateSystemGroup.cs
@@ -1,11 +1,0 @@
-﻿using Unity.Entities;
-
-namespace WaynGroup.Mgm.Ability
-{
-
-    [UpdateInGroup(typeof(AbilitySystemsGroup))]
-    public class AbilityUpdateSystemGroup : ComponentSystemGroup
-    {
-
-    }
-}
