--- Runtime/Systems/AbilityUpdateSystemGroup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d273ce7a821029c48885643ebc3a041e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
